Fix Absolute Links to Documentation Pages in Docstrings
GH Pages hosts the site at a `/linearmoney` subdirectory, so the absolute
links used in the in-source docstrings such as `/glossary.html` are not
working in the production documentation site.

This is difficult to automatically resolve since the api docs are built
by pdoc, so the url parsing features of mkdocs can't be used to expand
the absolute URL.

This is something to look into in the future since editing the absolute
URLs in the docstrings to match the site layout is not maintainable, but
for now, I have simply updated the source files with the correct
absolute links.