Remove Implicit Rounding from l10n Function
I originally had the l10n function round the passed in decimal amount before
formatting since I was thinking of formatting as a UI display feature, so
I assumed that users would always want a rounded value, but I realize now
that there are use cases for formatting a non-rounded value such as on
financial statements rounded to a certain precision.

Also, the implicit rounding behavior caused the interface of the `roundas`
function to bleed into the interface of the `l10n` function and complicated
the documentation making it more difficult to document and understand what
the `l10n` function was actually doing.

I have removed the implicit rounding from the `l10n` function, so now the
user has to provide a rounded value to the `l10n` function if they want the
value to be formatted with rounding. This allowed me to remove the `cash`
keyword argument from the `l10n` signature, and it makes the code more explicit
in what it is doing when using the `l10n` function.

I updated all of the documentation examples that use the `l10n` function including
the README and quickstart tutorial.

I also updated all of the related test cases to ensure that they are working
correctly with the new interface, and I added a new test case that checks that
localizing an unrounded value doesn't cause weird formatting like grouping
the digits after the decimal place.

While I was working on these changes, I also noticed a couple of small
tooling/documentation fixes I could make, so I updated the basic usage
example to better show the benefit of the library by using values in
different currencies in the example, and I added the README to the
files that get collected for doctests in the `doctests` tox
environment, so that any examples in the README will be checked for
correctness before being committed.