Prepare Release 0.1.0
I added a tox environment for testing against the built package after
publishing to TestPyPi and updated the README with more specific
information about the semver expectations prior to version 1.0.0.

I also removed the notice about not being able to create a PyPi package
since the restriction on new account creations was lifted today, so I am
now able to release to PyPi.