Replace CurrencySpace.axes with CurrencySpace.currencies in Quickstart
I originally used the CurrencySpace.axes property in the quickstart
tutorial for the valid options for <to_currency> and <from_currency>,
but these should have been CurrencySpace.currencies for faster
membership testing, so I have updated the quickstart tutorial to use
CurrencySpace.currencies instead of axes.

I also added a direct link to the quickstart tutorial at the top of the
README to make it more prominent for new users.