Add `system_locale` Function to `data` Module
When I was still working on this library in stealth mode, I had a
system_locale function that returned the locale of the running Python
process. This was used as a default locale because the original design
of the library used an OOP style and had a setable locale, so this was
required to ensure that the library had sane defaults for its internal
state, but after refactoring to make the library purely functional, this
was no longer required, so I removed the system_locale function as part
of a pruning pass where I removed most of the newly unneeded code.

I realized however that for local applications such as CLI or desktop
apps, we generally just want to use the system locale, and even though
this can be accomplished with the regular `locale` function by reading
the system locale using the stdlib's locale module, this is a major
annoyance for users, so I re-added the `system_locale` function as a
helper function to make this common use-case much simpler to implement
and make the library more approachable for simple use-cases.

I also updated the quickstart tutorial to use the new helper instead of
defining the system locale manually.

Also, while writing the test cases for the new `system_locale` function,
I realized that it would be helpful to have a simple way to access the
locale tag of a `LocaleData` instance as a property instead of
concatenating the `language` and `region` properties manually, so I
added a `tag` property that returns the locale tag as a string to the
`LocaleData` class.

Lastly, I added docstrings for the `id` property on the `LocaleData` and
`CurrencyData` classes to make the use of that property on Datasources
more clear.