Revise Linear Money Model Article
I revised the linear money model article for style and structure based
on feedback from a proofreader.

I had quite a few long/run-on sentences that I fixed by removing the
conjunctions with ", and" and simply starting a new sentence instead.

I also moved the Disclaimer section into the Abstract in order to set
the tone as constructive early on. I had originally added the Disclaimer
section since I felt like the Introduction came off as overly critical
of Fowler's pattern, which I didn't want to leave up to interpretation
since I have such high respect for Martin Fowler and his colleagues at
Thoughtworks. Unfortunately, this did break up the flow of the article
and felt a bit forced as well, so I moved the meaningful portion of the
Disclaimer into the Abstract and removed the Disclaimer section
completely.

I also removed the use of "just" in favor of more descriptive "only" or
"simply" where applicable based on feedback from the proofreader. I
originally assumed that it would be okay to use "just" since the overall
tone of the article is only semi-formal, but I definitely think it's
better without it. I did leave the use of "just" in a couple of places
where it's usage is in a very common sentence that should not cause
confusion.

I removed the use of "Okay, " at the beginning of sentences also based
on feedback from the proofreader... I'm honestly not sure what I was
thinking when I wrote those sentences. Don't write technical articles at
3am on a Sunday night.

I also added the original published date and a date for this update with
a link to the git history for the markdown file so that readers can
check the full history if they want to.

There are various other small style fixes as well, but overall the
content/meaning of the article has remained the same, so this is only a
style/structure revision.