Add `CLDR_VERSION` Package-level Member Variable
I added a `CLDR_VERSION` variable to the __init__.py for the linearmoney
package.

In order to have the CLDR_VERSION update automatically with the version
of the cldr data that is parsed into the library, I added a few lines
to the process_cldr_data.py script to open the
cldr-core package of the cldr-json data and read the `version` string.

I then write the version string to a new package resource file
`cldr_version.json`, which is loaded at import-time via the `resources`
module in order to ensure the CLDR_VERSION gives the correct version
information for the cldr data the library is currently using, even if
the user manually runs the processing script with a different version
of the cldr-json data than what linearmoney is distributed with and
rebuilds the package.

This should allow proper automated reporting of the correct cldr data
version without coupling the cldr version to the library source code.

This is not the most ideal solution as it requires a file read at
import-time, but it only happens once, and it is a very simple and
flexible implementation. A more robust solution that doesn't require
a package resource file read is likely possible, but it doesn't
provide much practical benefit over the simple package resource
solution, and I don't have the time to put into it right now, so I am
leaving this as is for now. This is something that can be improved in
the future though.