Add Tox Environments for Python 3.11 and 3.12
I added python version factors to tox.ini for python 3.11 and 3.12 on
the test environment to run the test suite against all supported
versions of Python.

The pyproject.toml file still gives Python 3.12 as the minimum version,
but the requirement for 3.12 will be dropped down to 3.11 in a separate
commit.

The 3.12 requirement was for the improved type alias and generics syntax
introduced in 3.12, but that has not been utilized yet, and even if it
is, it doesn't affect the runtime code, so we can use 3.11 as the
minimum version and document that type checking needs 3.12 if we do use
the 3.12 syntax at some point.

The different python versions need to be available for tox discovery, so
I will need to document the usage of pyenv with tox, but there isn't
much to it. Tox should use the available pyenv versions of Python as
long as they are installed thanks to the .python-version file.

While tweaking the tox config, I found a very embarrasing bug in the
mypy environment. I had originally called the mypy environment simply
"type", but at some point, I intended to rename it to "typecheck", but I
only changed the name in the `envlist` config variable, not the actual
environment. Because tox silently passes empty environments, I never
noticed that mypy wasn't being run when using `tox` without any
arguments. I have fixed this by renaming both the `envlist` name and the
actual environment name to `types`. There were a couple of simple type
errors in the `data` module that I fixed in this commit as well. I had
the wrong type in a few annotations. It was fixed with a simple rename
of the annotation types.