Reduce Required Python Version to 3.11
I changed the minimum Python version in pyproject.toml from 3.12 to
3.11.

I originally set the minimum Python version to 3.12 to take advantage of
the improved type alias and generics syntax in that version, but those
features have not been utilized yet, and they don't affect the runtime
code as long as we guard them in a `if TYPE_CHECKING` block, so we can
safely drop the minimum version to 3.11 for improved compatability.

We could drop the minimum version lower, but 3.11 was released in 2022
and it is the oldest bug fix version, so I think it's fine to only
support 3.11 and higher.

For reference, 3.9 will hit end of life in 2025, and 3.10 will be end of
life in 2026, but they are only receiving security patches at this
point, and most legacy projects are unlikely to adopt a new currency
library. Even if they did, the amount of work to port legacy code from a
bespoke solution or a different library to one with a unique data model
like linearmoney would probably be more than the work required to
upgrade to a newer version of Python.

See https://devguide.python.org/versions/ for the official support
lifetimes of all current Python versions.