Remove `makedocs` Environment from Tox Default Envlist
Currently, running tox without arguments regenerates the documentation
site, which is desirable while working on local development, but it
causes problems with VCS since the sitemap and updated timestamps are
changed, so it creates a huge diff even if nothing really changed in the
documentation content.

I removed the `makedocs` environment from the default `envlist`, so that
it won't run when running `tox` without arguments. This should prevent a
bunch of unnecessary changes cluttering the diff when working on bug
fixes or features.

Unfortunately, this also means that we will now have to run `tox run -e
makedocs` manually when we actually do want to update the documentation
site, but this can be done in a CI job once the repo has a CI setup, so
that we ensure that merging a release into main always gives the correct
documentation for the current version.

If a developer wants to view the documentation site locally, they will
have to run the `makedocs` environment manually.