Change Package Build Backend to Hatchling
This is the first step of moving the project from Poetry + Pyenv + Tox
to Hatch for all dependency and venv management.

I changed the build backend and all metadata in pyproject.toml to use
Hatchling. I have not added any of the test environments yet. That work
will be done in a separate commit dedicated to replacing Tox.