Replace Tox with Hatch Environments
I added hatch env configs to the pyproject.toml file and removed the
dependency on tox as well as tox.ini. I also removed the poetry.toml and
poetry.lock files as this change removes the dependence on Poetry as
well.

With these changes, the project now relies on Hatch for its testing,
venv management, and buildsystem, but this removes the dependencies on
Poetry, Tox, and Pyenv for project management all at once.

Hatch has pretty simple venv management, and it doesn't add all the
complexity that Poetry layers on top of the pyproject.toml build
declarations. Poetry felt like using npm, which gives me
nightmares after working with Javascript professionally, but Hatch seems
to be very Pythonic in how it works so far.

Also, Hatch automatically installs missing python interpreters based on the
versions declared in the defined environments, so we don't have to
document how to install and maintain separate versions of Python in
order for contributors to run the test suite.

This is a big win for workflow simplicity, and it should make
documenting contribution guidelines easier.

The only limitation that I see with Hatch at the moment is that it
doesn't support any kind of lock file for dependencies. I believe that
the Hatch maintainers are working on a solution for this, so it should
have feature parity with Poetry in that regard eventually. In any case,
lockfiles aren't a big deal for this project since linearmoney has no
production dependencies, but resolution of incompatible dev dependencies
could eventually become a headache. If that ever happens, we can always
just use `pip freeze` as a workaround until Hatch adds lockfile support.