Add GitHub Actions CI
I added a github actions workflow for running the test suite, type
checks, and the linter on PRs and direct commits to both dev and main.

The CI test both Linux and MacOS, but Windows is commented out of the
workflow at the moment. This is because Windows locales don't use the
posix format for their short strings, so the locale system is broken on
Windows. I don't see any reason to waste CI minutes on every run while
Windows is unsupported.

Also, I may decide not to support Windows at all since this library is
mainly intended for web servers, and very few people are hosting
websites on Windows.

In any case, there is a Windows job in the workflow file, but it is
commented out for now.

Additionally, the Linux job is currently failing because I did not
consider the "C"/"POSIX" locale when designing the locale data. This is
an actual bug though, so I am leaving the CI workflow as is, and it
should run successfully once the bug with the C locale is fixed.

Lastly, I updated the Hatch run commands in the pyproject.toml file to
work well with both local development and a hosted CI environment.

I changed the `ci` command to propogate the exit codes from its steps. I
also removed the `format` step from the `ci` command to avoid a
false-pass on the linter steps in the CI. I added a new `all` command
that runs the autoformatter first and fails if there are any errors in
the formatter run but then runs all of the other hatch commands without
propogating exit codes so that the developer can see all errors at once
when developing locally.