Interpret POSIX C Locale as en_US
I changed the `system_locale` function to check if the locale is the
default POSIX C locale and interpret it as `en_US` if that is the case.

The C locale is close to en_US, so this is a reasonable mapping to a
valid linearmoney internal locale.

This behavior may change while working on #15, but for now, this should
be a sufficient workaround for this bug.

Most workstations won't be using the C locale, but it's the default on
POSIX systems, so machines like an Ubuntu server could have this locale
if they weren't configured for a specific locale before partitioning.

I also added a test case for the C locale and modified the other test
cases for the `system_locale` function based on the new behavior.

I also had to modify the GH Actions workflow file to make sure all
environments have the requisite locales.

Fixes #14