Update Build from Source Instructions in README.
I updated the instructions for building from source to use regular
`python -m build` instead of `poetry` since we dropped poetry in favor
of hatch, and hatch is only needed to run the tests, not build the
project as an end-user.

I also added a link to the issue tracking the work on the
locale/currency data redesign to the roadmap section.