Add Contributing Guidelines
I added a basic CONTRIBUTING.md file so that the project can begin
accepting public contributions if anyone is interested in helping out.

I also updated the Contributing section of the readme and checked off
the related milestone in the roadmap.

I did not choose a Code of Conduct as I want to put more time and
thought into that kind of decision, but I did include a *Community
Interaction* section in the contributing guide that gives a formal
statement that professionalism is expected.

I doubt this project will need any kind of formal community guidelines,
but it doesn't hurt to have a statement in place that can be referenced
if needed.