Add Automated Testing for CLDR Data Processing Script
I added a `tests/cldr.py` test script for regression testing the cldr
data processing script when making changes.

The test script simply checks the produced JSON files against files
produced by a known correct version of the script. Basically, it's just
a snapshot test.

This means that the reference JSON files in the tests directory need to be
updated when the version of the CLDR data is updated, but I have the
`cldr_version.json` file as part of the test data, so if the version is
accidentally updated, then the test suite will fail on the check of that
file, and we will know that we need to either roll back changes or
verify that the new version works as expected and update the reference
JSON.