Fix Regression in CLDR Processing Script
When I wrapped the loading of locale data from the CLDR json files into
memory in a function, I forgot to call the function afterwards, so the
cldr data processing script was not creating any locale data.

This is fixed.

I also added a new `cldr` environment to Hatch for building and testing
the cldr data. The tests are working correctly. That's actually how I
noticed that the script was broken and where to look for the problem.