Add Contributing Instructions for Updating CLDR Data
I added a section to the CONTRIBUTING.md file explaining the process for
updating the CLDR data used by the library.

I also changed the structure of the cldr test data to keep the test
script and all data in a dedicated `tests/cldr` directory instead of
having everything sitting in the top-level `tests` directory.