Refactor CLDR Data Processing Script
fix #11

I finished the primary refactoring the `process_cldr_data.py` script.

It's still not great, but it's broken up into more clearly delineated
function now, so it should be much easier to make changes to the way we
parse the data or to change the output format.

Through this refactoring process, I also found a couple bugs in the
script that were causing the incorrect data to be parsed, so this commit
also includes updated currencies and locales data.

Specifically, several currencies were being parsed as having
`cash_places = 2` when they should have been `cash_places = 0`. This was
caused by using `j` instead of `i` in the loop in the old
`_build_fractions_data` function. This caused the branch that read the
data for `cash_places` to get skipped for all currencies, so any that
had specific `cash_places` would instead get the value of the `DEFAULT`
currency.

The other problem was that the comparisons with `" "` weren't working
for locales that use some other unicode whitespace character for their
spaces in the formatting patterns. This caused several locales to have
incorrect currency symbol and positive/negative sign placement.

I added a helper function for normalizing whitespace with the
`unicodedata` module, and this is now fixed. The new data should have
the correct sign and symbol placement and spacing.