Fix Failure in Localizations Test
The updates to the locales data affected the fr_FR locale, which was
being parsed incorrectly before and not including the space between the
numeric portion and the currency symbol.

The changes to the cldr data processing script fix this error in the
parsed locales data, but I forgot to update the test case that checkes
for non-ascii characters in the `scalars.l10n` function. This was not
caught because the venv that hatch created did not update with the new
data files when running the test suite locally, so it was not caught
until CI ran on the PR for this.

I manually recreated the venv locally to confirm the changes fix the
failing test case and reflect the correct localization formatting.