Add SQLAlchemy Integrations
I have added an `ext` namespace package to the library.

This package currently contains a single module `sqlalchemy.py` that
provides sqlalchemy column types for storing MoneyVectors in a sql
database.

The `ext` namespace package is distributed with the regular package, but
it is not imported in the main `__init__.py` file, so it doesn't cause
any import issues if the user hasn't installed the optional
`linearmoney[sqlalchemy]` dependencies.

I have added two column types `VectorMoney` and `AtomicMoney`.
VectorMoney stores the MoneyVector as a serialized string, so it
preserves all information including the currency space. AtomicMoney
stores the MoneyVector as an integer of the smallest denomination of
whatever currency is provided to the column constructor on model
declaration. For this reason, AtomicMoney is intended to only be used by
single-currency applcations that want to take advantage of the more
efficient storage format and the ability to use aggregate functions
inside the db. Multi-currency applications should prefer VectorMoney.

I have added some basic test cases to the new
`tests/ext/sqlalchemy_test.py` test file, but they are not
comprehensive. They test the basic serialization correctness, which is
the most important thing, but they don't exercise any of the expected
error conditions. Specifically, I need to add tests for SpaceError when
attempting to store a MoneyVector with a different currency space in an
AtomicMoney column.

One last thing that needs to get taken care of before this feature can
be merged into dev is to get mypy passing. Currently, mypy is failing
because it can't find type definitions for sqlalchemy. I need to look
into this further to figure out why mypy can't resolve the types for
sqlalchemy. This branch can't be merged until this is sorted out.