Fix Static Typing in SQLAlchemy Extension
Mypy isn't able to find types for sqlalchemy unless it is installed
with "sqlalchemy[mypy]". I have added this as a dependency to the
`types` hatch environment.

I also had to add a couple of type: ignore comments to the sqlalchemy
integrations to get better type checking.

SQLAlchemy types the `value` argument to the `process_bind_param` and
`process_result_value` methods of the `TypeDecorator` class as `Any |
None`, which makes sense for methods that need to be overriden with any
value, but it essentially disables static type checking on those
methods.

We can add a type annotation for the specific types we want to support
in the override, but mypy complains about Liskov since this is
technically a violation of type theory (subtypes can only extend their
parent types). Violating Liskov doesn't make any difference in this case
since the supertype is untyped, so I have added `type: ignore[override]`
comments to these lines. This should ignore the Liskov error but still
provide type checking against the provided annotations.

This could potentially cause a typing problem if someone tries to
subclass `VectorMoney` or `AtomicMoney` for some reason. That's what the
Liskov Substitution Principle is intended to address, but subclassing
those types would be stupid, so I'm not worried about it.