Add Test Cases for Math Errors in AtomicMoney Type
I added a couple test cases that make sure the AtomicMoney column is
raising errors when passed a vector that will break the math model.

The AtomicMoney column uses a single-currency space to evaluate the
stored vector, so SpaceError gets raised if the user tries to store
any vectors of more than one dimension in this column. This will be
caught by sqlalchemy and reraised as a `sqlalchemy.exc.StatementError`.

I also fixed a bug in the test suite for the sqlalchemy extension.
The test cases for the AtomicMoney column were using the VectorMoney
column in the test code, so the AtomicMoney type actually wasn't being
exercised in the test suite. This is because I copy/pasted the test
functions for the VectorMoney type and changed the setup, but I forgot
to update the model class used in the sqlalchemy queries, which is a
very facepalmy thing to do, but stuff happens. This is why we write
lots of tests. I realized this when I added test cases that expected
an error to be raised by the AtomicMoney column that didn't get raised
because it was using the VectorMoney column instead. :)

These test cases could be parametrized, but there's enough difference
in their setup that it's not worth it for only two inputs.

All tests are passing and testing the correct classes. Coverage
report is also showing 100% coverage of both classes.