Fix Doctest Examples in SQLAlchemy Extension
The docstrings for VectorMoney and AtomicMoney in the sqlalchemy
integrations were not formatted correctly, so this was causing failures
in the doctests. I have updated the docstrings to use the correct
formatting.

I also added the `linearmoney.ext` module to the modules collected in
the `documentation/run_doctests.py` script since that script explicitly
collects docstring examples for packages, so it wasn't running the
doctests for the namespace package before.