Add Documentation Generation for SQLAlchemy Extension
I modified the module structure a bit to allow pdoc to find the
`linearmoney.ext.sqlalchemy` module and generate api documentation for
it.

There were a couple of ways to make this work. One way was to simply
pass in the namespace package in the call to `pdoc` on the commandline.
Another way was to make the `ext` package a regular python package with
an `__init__.py` file and let pdoc find and document everything based on
the `__all__` values just like the rest of the framework. I chose the
latter because there wasn't any reason to use a namespace package in the
first place, and having the `__init__.py` file allows me to document the
package itself as well as its contents, which is helpful for the api
doc site.

I also added some improvements to the docstrings in the sqlalchemy
module, and I made the `currency`, `forex`, and `space` class vars
read-only properties instead of mutable class members. This makes the
intended usage contractual just like the classes in the core library.

Lastly, I fixed a couple of typos in the docstrings of the
`linearmoney.vector` module. I was using `Munny` instead of `Money`.
This is a remnant of the library when I was still working on it in
stealth mode.

I did build the docs in this commit because I wanted to keep the source
changes visible in the commit diff. I built the docs locally and
verified that everything is generating correctly, so I don't anticipate
any problems, but I will double-check everything when I build the
documentation for the next release.