Fix SQAlchemy Integrations on Nullable Columns
There was a bug in the sqlalchemy integrations that caused the custom
column types to raise when used on nullable columns.

This was caused by the custom column types not checking for `None` in
the `process_bind_param` and `process_result_value` functions, which is
a silly noob mistake, but it's something that I completely overlooked
when I first added the column types.

This is fixed now, and I have added a couple of simple regression test
cases for it as well.