Remove Redundant Pathspec from Doctest Collection
The `run_doctests.py` script was manually collecting the
`linearmoney.ext` module, but the extensions module is collected as part
of the root `linearmoney` module by pdoc, so this was redundant and was
causing a warning to be printed when running the `docs:test-in-source`
environment script.

This is fixed now.