Remove Extraneous Rounding Test Cases
The parametrized regression test for values with leading zeroes in the
fractional component had duplicate test cases.

The test cases were named with a `_round_up` or `_round_down` suffix.
This is a remnant of a previous version of the library where I was
managing the `decimal.Context` inside the library and was testing
different rounding modes. I originally included the context management
because I wanted the library to be easily usable for people without any
knowledge of fixed-point decimal. I eventually realized that anyone
programming a sufficiently complicated application that needs the
features of this library will have to grapple with decimal at some
point, so I removed the Context management before the first public release.

IIRC, after the above change, I simply removed the `rounding_mode`
parameter from these test cases without changing anything else, but
that parameter was the only variance between the `_round_up` and
`_round_down` cases, so this resulted in duplicate test cases where the
rounding mode was parametrized.