[BREAKING] Redesign Rounding Algorithms and Functions
The rounding behavior was inconsistent. I don't know why I designed the
rounding functions to round to the nearest denomination in the
fractional portion of a monetary amount independently of the whole
dollar portion. This causes problems with subsequent calculations since
it creates a monetary value that is not divisible by its smallest
denomination.

To clarify, let's use the example of denomination=15 and places=2.

Given a value of 10.40, the original `scalar.roundas` function would
round this to `10.45`. This made sense to me when I first implemented
this since 45 is divisible by 15, but thinking about this a bit more,
1045 is not divisible by 15, which means that subsequent calculations
using this value will likely result in some kind of cumulative rounding.
There may be applications that need to calculate the denominational
rounding of the fractional portion separately, so I thought about adding
a `piecewise` rounding function to make this distinction explicit, but
this is likely a rare edge case, so instead, I simply added an example
piecewise rounding algorithm to the docstring for the rounding function.

I rewrote the internals of the rounding functions so that they would use
the atomic value of the monetary amount as the source of truth for
rounding behavior. This means that denominational rounding should always
result in a value that satisfies the following invariant:
`atomic(value, currency) % currency.data["denomination"] == 0`

In other words, the rounded value should be divisible by the
denomination when taken as an integral value of its fractional currency.

Removing the piecewise rounding behavior is a breaking change, but it
makes the behavior consistent and predictable, and it also simplifies
the implementation significantly.

In addition to changing the behavior of the rounding algorithm, I also
refactored the rounding functions with the following changes:
- rename `scalar.round_as` -> `round.as_currency`
- rename `scalar.round_to` -> `round.to_places`
- rename `scalar.atomic` -> `round.atomic`
- add `round.to_nearest`

The new `round.to_nearest` function rounds the given decimal `amount` to
the nearest multiple of the integer `value`. This is useful for a lot of
financial applications where amounts need to be displayed in thousands
or millions such as on financial statements.

I also removed a lot of the private functions that were acting as steps
in the rounding algorithm and moved things around so that the public
functions would be used instead. The rounding algorithm is detailed in
a comment at the top of the `round` module.

Another change that I made was that I removed the implicit behavior of
making the `places` match the digits in `denomination` when
`denomination` has more digits than `places` and made that situation an
error instead. I also moved the detection for this situation into the
`data.currency` function so that users are not able to create a
`CurrencyData` instance that contains this invalid combination. I added
a new exception type `InvalidDataError` that gets raised when attempting
to call the `currency` function with this invalid combination. I also
plan to use this exception type for locale data if I come across any
incompatible combinations of formatting data when I redesign the
localization system later on.

This commit also includes .ignore, .fdignore, and .rdignore files.
These files are for my own development workflow. They allow me to ignore
files from neovim's fuzzy file search without excluding them from git
tracking via .gitignore.